Refactor methods to enum extensions
